% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_NCBI_BLAST_XML_results_into_datatable.R
\name{aggregate_NCBI_BLAST_XMLs2DT}
\alias{aggregate_NCBI_BLAST_XMLs2DT}
\title{Aggregates NCBI BLAST results from multiple XML files into an R data.table.}
\usage{
aggregate_NCBI_BLAST_XMLs2DT(dir.to.xmls, ncores = 1)
}
\arguments{
\item{dir.to.xmls}{A \code{character} specifying a directory to be searched
recursively for NCBI BLAST XML results (For more
information about XML results from NCBI BLAST see
\link{NCBI_BLAST_XML2DT}).}

\item{ncores}{An \code{integer} specifying the number of cores or
threads to be used for parallel processing.}
}
\value{
A \code{data.table} aggregating all NCBI BLAST results from all XML
        files found.
}
\description{
Aggregates NCBI BLAST results from multiple XML files into an R data.table.
}
\examples{
#Aggregate my NCBI BLAST results using 2 cores into an R data.table 
aggregate_NCBI_BLAST_XMLs2DT(dir.to.xmls = "my_dir/with/NCBI_BLAST_results/",
  ncores = 2)
}
\references{
\itemize{
 \item{\href{https://www.ncbi.nlm.nih.gov/BLAST/tutorial/Altschul-1.html}{The Statistics of Sequence Similarity Scores.}}
 \item{\href{https://cran.r-project.org/web/packages/hoardeR/index.html}{hoardeR: Collect and Retrieve Annotation Data for Various Genomic Data Using Different Webservices.}}
 \item{\href{https://cran.r-project.org/web/packages/xml2/index.html}{xml2: Parse XML.}}
}
}
\author{
Yoann Pageaud.
}
