% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_NCBI_BLAST_request.R
\name{get.NCBI.BLAST2DT}
\alias{get.NCBI.BLAST2DT}
\title{Automatically submits all sequences and retrieves all results in a
data.table.}
\usage{
get.NCBI.BLAST2DT(
  sequences,
  db = "genomic/9606/GCF_000001405.39",
  res.dir,
  ncores,
  auto.rm.dir = TRUE,
  delay.req = 10,
  email
)
}
\arguments{
\item{sequences}{Either a \code{list} of DNA sequences, or a
\code{data.frame} in a BED-like format containing Genbank
accession IDs and coordinates to extract sequences from.
\cr For more information about the data.frame format see
'GBaccess.bed' argument in \link{prepare.gb.access}.
\cr For more information about the list format see the
example in \link{submit_NCBI_BLAST}.}

\item{db}{A \code{character} string specifying an NCBI genome or
reference sequence set database on which the submitted
sequences will be BLASTed (Default using GRCh38 (hg38)
genome assembly database:
db = "genomic/9606/GCF_000001405.39").
For more supported databases see Details section in
\link{submit_NCBI_BLAST}.}

\item{res.dir}{A \code{character} string specifying the directory where
NCBI BLAST submission results will be stored. Each BLAST
result is stored in a separate folder within the given
directory. Blast hits for each submission are available as
a XML file in each matching folder.}

\item{ncores}{An \code{integer} to specify the number of cores/threads
to be used to parallel-compute probes intensities.}

\item{auto.rm.dir}{A \code{logical} specifying whether the results directory
should be immediately deleted after function execution end
(auto.rm.dir = TRUE) or not (auto.rm.dir = FALSE). If
FALSE, the folder will remain available after execution
with all raw results in it.}

\item{delay.req}{Mandatory seconds between BLAST request submissions as an
\code{integer} (Default: delay.req = 10).}

\item{email}{Mandatory E-mail adress for NCBI BLAST request, specified
as a \code{character} string.}
}
\value{
A \code{data.table} aggregating all NCBI BLAST results.
}
\description{
Automatically submits all sequences and retrieves all results in a
data.table.
}
\examples{
#Create 1 row data.frame for the example
df <- data.frame("GBaccessID" = "AC073318", "Start" = 71401, "End" = 72401)
#Submit the sequence extracted from AC073318 to NCBI BLAST API
example.dt <- get.NCBI.BLAST2DT(
  sequences = df, db = "genomic/9606/GCF_000001405.25",
  res.dir = "~/result_BLAST", ncores = 2, auto.rm.dir = FALSE,
  email = "myemailadress@dkfz.de")
#example.dt is a data.table containing all BLAST hits for the sequence submitted.
}
\references{
\itemize{
 \item{\href{https://cran.r-project.org/web/packages/hoardeR/index.html}{hoardeR: Collect and Retrieve Annotation Data for Various Genomic Data Using Different Webservices.}}
 \item{\href{https://academic.oup.com/bioinformatics/article/35/3/526/5055127}{Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R. Bioinformatics 35: 526–528. doi:10.1093/bioinformatics/bty633. HAL: ird-01920132.}}
 \item{\href{https://www.ncbi.nlm.nih.gov/BLAST/tutorial/Altschul-1.html}{The Statistics of Sequence Similarity Scores.}}
 \item{\href{https://academic.oup.com/nar/article/36/suppl_2/W5/2505810}{Johnson, M. et al. NCBI BLAST: a better web interface. Nucleic Acids Research 36, W5–W9 (2008).}}
}
}
\author{
Yoann Pageaud.
}
