% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_NCBI_BLAST_request.R
\name{prepare.gb.access}
\alias{prepare.gb.access}
\title{Prepares a set of GenBank Accession IDs for NCBI BLAST submission}
\usage{
prepare.gb.access(GBaccess.bed, ncores = 1)
}
\arguments{
\item{GBaccess.bed}{A \code{data.table} or a \code{data.frame} in a BED-like
format containing columns as following:
\itemize{
 \item{column 1: The Genbank accession IDs}
 \item{column 2: The start position of the sequence to
 keep}
 \item{column 3: The end position of sequence to keep}
}}

\item{ncores}{An \code{integer} specifying the number of cores or
threads to be used for parallel processing.}
}
\value{
A \code{list} of sequences based on information provided in GBaccess.
}
\description{
Prepares a set of GenBank Accession IDs for NCBI BLAST submission
}
\examples{
#Create example 1 row data.frame
df <- data.frame("GBaccessID" = "AC073318", "Start" = 71401, "End" = 120576)
#Prepare sequence for submission to NCBI BLAST
seq.list <- prepare.gb.access(GBaccess.bed = df)
# seq.list can now be used by submit_NCBI_BLAST
}
\references{
\href{https://academic.oup.com/bioinformatics/article/35/3/526/5055127}{Paradis E. & Schliep K. 2019. ape 5.0: an environment for modern phylogenetics and evolutionary analyses in R. Bioinformatics 35: 526–528. doi:10.1093/bioinformatics/bty633. HAL: ird-01920132.}
}
\author{
Yoann Pageaud.
}
