% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_NCBI_BLAST_request.R
\name{split.queries.df}
\alias{split.queries.df}
\title{Splits a data.frame of Genbank IDs and intervals into a data.table of smaller
intervals.}
\usage{
\method{split}{queries.df}(df, by = 7000)
}
\arguments{
\item{df}{A \code{data.frame} or a \code{data.table} in a BED-like format
following the same specifications as 'GBaccess.bed' in
\link{prepare.gb.access}.}

\item{by}{An \code{integer} to specify the length of the output DNA sequences
you wish your queries to be splitted into, in base pair
(Default: by = 7000).}
}
\value{
A \code{data.table} of the Genbank IDs with intervals of the
        specified length. You can then subset the intervals of interest
        before submitting the data.table to \link{prepare.gb.access} or
        \link{get.NCBI.BLAST2DT}.
}
\description{
Splits a data.frame of Genbank IDs and intervals into a data.table of smaller
intervals.
}
\examples{
#Using an example data.frame of 1 Genbank ID
split.queries.df(
  df = data.frame("GB.access" = "AC073318", "Start" = 71401, "End" = 120576),
  by = 7025)
}
\author{
Yoann Pageaud.
}
