% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_NCBI_BLAST_request.R
\name{split_queries}
\alias{split_queries}
\title{Splits a set of BLAST queries into a list of smaller DNA sequences.}
\usage{
split_queries(x = x, by = 7000, ncores = 1)
}
\arguments{
\item{x}{Supports 2 possible formats:
\itemize{
 \item{A \code{data.frame} in a BED-like format following the
 same specifications as 'GBaccess.bed' in
 \link{prepare.gb.access}.}
 \item{A \code{list} of DNA sequences.}
}}

\item{by}{An \code{integer} to specify the length of the output DNA
sequences you wish your queries to be splitted into, in base
pair (Default: by = 7000).}

\item{ncores}{An \code{integer} specifying the number of cores or threads to
be used for parallel processing.}
}
\value{
A \code{list} of DNA sequences of the requested length.
}
\description{
Splits a set of BLAST queries into a list of smaller DNA sequences.
}
\examples{
#Using an example data.frame of 1 Genbank ID
ls.seq <- split_queries(
  x = data.frame("GB.access" = "AC073318", "Start" = 71401, "End" = 120576),
  by = 7025)
#Using an example list of DNA sequences
ls.seq <- list(
  "7qtel" = "CCCTAACACTGTTAGGGTTATTATGTTGACTGTTCTCATTGCTGTCTTAG",
  "1ptel" = "GATCCTTGAAGCGCCCCCAAGGGCATCTTCTCAAAGTTGGATGTGTGCAT",
  "17qtel" = "CCCTAACCCTAAACCCTAGCCCTAGCCCTAGCCCTAGCCCTAGCCCTAGC")
ls.seq <- split_queries(x = ls.seq, by = 10)
}
\author{
Yoann Pageaud.
}
