% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit_NCBI_BLAST_request.R
\name{submit_NCBI_BLAST}
\alias{submit_NCBI_BLAST}
\title{Submits a list of DNA sequences to NCBI BLAST for alignment to a sequence
database.}
\usage{
submit_NCBI_BLAST(
  seq.list,
  res.dir,
  delay.req = 10,
  email,
  db = "genomic/9606/GCF_000001405.25"
)
}
\arguments{
\item{seq.list}{A named \code{list} of DNA sequences stored as character
strings to be BLASTed on NCBI servers databases.}

\item{res.dir}{A \code{character} string specifying the directory where
NCBI BLAST submission results will be stored. Each BLAST
result is stored in a separate folder within the given
directory. Blast hits for each submission are available as a
XML file in each matching folder.}

\item{delay.req}{Mandatory seconds between BLAST request submissions as an
\code{integer} (Default: delay.req = 10).}

\item{email}{Mandatory E-mail adress for NCBI BLAST request, specifiedas
as a \code{character} string.}

\item{db}{A \code{character} string specifying an NCBI genome or
reference sequence set database on which the submitted
sequences will be BLASTed (Default using GRCh37/hg19 genome
assembly database: db = "genomic/9606/GCF_000001405.25").}
}
\value{
A \code{type} object returned description.
}
\description{
Submits a list of DNA sequences to NCBI BLAST for alignment to a sequence
database.
}
\examples{
#Create example list of sequences you wish to BLAST using NCBI BLAST API.
ls.seq <- list(
  "7qtel" = "CCCTAACACTGTTAGGGTTATTATGTTGACTGTTCTCATTGCTGTCTTAG",
  "1ptel" = "GATCCTTGAAGCGCCCCCAAGGGCATCTTCTCAAAGTTGGATGTGTGCAT",
  "17qtel" = "CCCTAACCCTAAACCCTAGCCCTAGCCCTAGCCCTAGCCCTAGCCCTAGC")
#Submit the list of sequences to NCBI BLAST using the GRCh37/hg19 genome assembly database.
submit_NCBI_BLAST(
  seq.list = ls.seq, res.dir = "~/result_directory",
  email = "myemailadress@dkfz.de", db = "genomic/9606/GCF_000001405.25")
}
\references{

}
\author{
Yoann Pageaud.
}
